Feature improve tracking (#54)
* Updated styling for the back button

* Fixed start button wrong location

* Further improved startbutton positioning

* Extra margin top for position button

* New component <HoverWithFriends />

* Added <TrackinInformation /> component

* Removed unnecessary styling

* Fixed achievement name

* Fix require warning

* Tracking improvements to trackingprocider

* Improved achievement and activity feed card

* Improved feed cards

* Added check for same geofence id on start/stop activity

* Removed unnecessary (for now at least) background location tasks

* Added check for exiting hover zone when tracking. If so pause score updates, but keep location updates

* Updated location interval

* Minor improvements

* Added new Keyboardavoider

* Update keyboardavoider

* Update tracking info positioning

* Lint fix

* Added start tracking when reentering hover zone

* Fixed (?) keyboard avoidance issue for Android as well

* Fixed and issue with the logic in HoverNavigator

* Simplified challenge feed card

* Updated challenge feed card UI

* Update ChallengeFeedCard.tsx

* Improved tracking

* Fixed auto start/stop tracking provider

* fixed issue where tracking view was'nt touchable
