+++ components/feed/AchievementFeedCard.tsx
@@ -14,19 +14,24 @@ interface AchievementFeedCardProps {
 const AchievementFeedCard: React.FC<AchievementFeedCardProps> = ({ data }: AchievementFeedCardProps) => {
   return (
     <View style={styles.card}>
-      <TouchableProfile user_id={data.user.id} name={data.user.name}>
-        <View style={styles.topBar}>
-          <Image
-            source={{ uri: data.user.picture ? data.user.picture : defaultUserProfile.picture }}
-            style={styles.avatar}
-          />
-          <Text style={styles.nameText}>{data.user.name}</Text>
-        </View>
-      </TouchableProfile>
       <View style={styles.main}>
-        <View style={styles.description}>
-          <Text style={styles.descriptionText}>{data.achievement.description}</Text>
+        <View style={styles.infoContainer}>
+          <TouchableProfile user_id={data.user.id} name={data.user.name}>
+            <View style={styles.topBar}>
+              <Image
+                source={{ uri: data.user.picture ? data.user.picture : defaultUserProfile.picture }}
+                style={styles.avatar}
+              />
+              <Text style={styles.nameText} numberOfLines={1}>
+                {data.user.name}
+              </Text>
+            </View>
+          </TouchableProfile>
+          <View style={styles.description}>
+            <Text style={styles.descriptionText}>{data.achievement.description}</Text>
+          </View>
         </View>
+
         <View style={styles.achievement}>
           <Achievement achievement={data.achievement} />
         </View>
@@ -49,30 +54,35 @@ const styles = StyleSheet.create({
     flexDirection: 'row',
     justifyContent: 'flex-start',
   },
+  infoContainer: {
+    width: '68%',
+  },
   nameText: {
     ...Typography.headerText,
     fontSize: 20,
-    lineHeight: 50,
+    lineHeight: 30,
+    flexWrap: 'wrap',
+    marginTop: Spacing.smaller,
+    width: '80%',
   },
   description: {
     flexDirection: 'row',
     justifyContent: 'flex-start',
     alignItems: 'center',
-    width: '65%',
+    marginTop: Spacing.smaller,
   },
   descriptionText: {
     ...Typography.headerText,
-    fontSize: 24,
+    fontSize: 20,
     flexWrap: 'wrap',
     flexDirection: 'column',
-    maxWidth: '100%',
     paddingLeft: Spacing.smaller,
+    width: '100%',
   },
   main: {
     flexDirection: 'row',
     justifyContent: 'space-between',
     marginBottom: Spacing.smaller,
-    width: '100%',
   },
   footer: {
     flexDirection: 'row',
@@ -90,7 +100,6 @@ const styles = StyleSheet.create({
     marginRight: Spacing.small,
   },
   achievement: {
-    marginTop: -40,
     width: '35%',
   },
 });
