+++ components/providers/TrackingProvider.tsx
@@ -2,7 +2,6 @@ import React, { useState, ReactNode, useEffect } from 'react';
 import { GeoFence } from '../../types/geoFenceTypes';
 import { convertToGeoFences } from '../../helpers/objectMappers';
 import { usePermissions, LOCATION, PermissionResponse } from 'expo-permissions';
-import { startBackgroundUpdate, stopBackgroundUpdate } from '../../tasks/locationBackgroundTasks';
 import { useGeofencesQuery } from '../../graphql/queries/Geofences.generated';
 import { useInsertActivityMutation } from '../../graphql/mutations/InsertActivity.generated';
 import { Alert } from 'react-native';
@@ -15,14 +14,15 @@ import useAuthentication from '../../hooks/useAuthentication';
 import { LatLng } from 'react-native-maps';
 import { Activities_Insert_Input } from '../../types/types';
 
-interface Props {
-  children: ReactNode;
-}
-
 export enum TrackingState {
   EXPLORE,
   TRACKING,
+  TRACKINGPAUSED,
   PUBLISH,
+}
+
+interface Props {
+  children: ReactNode;
 }
 
 interface TrackingContextValues {
@@ -72,6 +72,7 @@ export const TrackingContext = React.createContext<TrackingContextValues>({
   friendId: '',
   setFriendId: () => console.error('Function not initialized'),
 });
+
 TrackingContext.displayName = 'TrackingContext';
 
 export const TrackingProvider = ({ children }: Props) => {
@@ -82,6 +83,7 @@ export const TrackingProvider = ({ children }: Props) => {
   const [geoFences, setGeoFences] = useState<GeoFence[]>([]);
   const [unUploadedActivities, setUnUploadedActivities] = useState<Activities_Insert_Input[]>([]);
   const [insideGeoFence, setInsideGeoFence] = useState<GeoFence | null>(null);
+  const [startGeoFence, setStartGeofence] = useState<GeoFence | null>(null);
   const [trackingState, setTrackingState] = useState<TrackingState>(TrackingState.EXPLORE);
   const [trackingStart, setTrackingStart] = useState('');
   const [score, setScore] = useState(0);
@@ -100,31 +102,19 @@ export const TrackingProvider = ({ children }: Props) => {
     if (geoFenceFetchError) console.error(geoFenceFetchError.message);
   }, [geoFenceData, geoFenceFetchError]);
 
-  useEffect(() => {
-    if (locationPermission && locationPermission.status === 'granted') {
-      console.log('Start background update');
-      startBackgroundUpdate();
-    } else if (locationPermission && locationPermission.status !== 'granted') {
-      console.log('Stopped background update. Permission not accepted');
-      stopBackgroundUpdate();
-    }
-  }, [locationPermission]);
-
   const externalUpdateUserLocation = (newUserLocation: LatLng) => {
-    if (userLocation === null) {
-      updateUserLocation({
-        coords: {
-          latitude: newUserLocation.latitude,
-          longitude: newUserLocation.longitude,
-          altitude: null,
-          accuracy: null,
-          altitudeAccuracy: null,
-          heading: null,
-          speed: null,
-        },
-        timestamp: Date.now(),
-      });
-    }
+    updateUserLocation({
+      coords: {
+        latitude: newUserLocation.latitude,
+        longitude: newUserLocation.longitude,
+        altitude: null,
+        accuracy: null,
+        altitudeAccuracy: null,
+        heading: null,
+        speed: null,
+      },
+      timestamp: Date.now(),
+    });
   };
 
   const updateUserLocation = (newUserLocation: LocationObject) => {
@@ -135,38 +125,57 @@ export const TrackingProvider = ({ children }: Props) => {
     );
 
     const insideGeoFenceCheck = insideGeoFences(newUserLocation, geoFences);
+
     if (insideGeoFenceCheck) {
-      setInsideGeoFence(insideGeoFenceCheck);
       console.log('Inside geo fence!');
+      setInsideGeoFence(insideGeoFenceCheck);
+      if (trackingState === TrackingState.TRACKINGPAUSED) {
+        setTrackingState(TrackingState.TRACKING);
+      }
     } else {
       console.log('Outside geofence!');
+      setInsideGeoFence(null);
+      if (trackingState === TrackingState.TRACKING) {
+        setTrackingState(TrackingState.TRACKINGPAUSED);
+        // TODO: Insert push notification
+      }
     }
   };
 
-  // Update user location every 5 seconds if tracking
-  useInterval(
-    async () => {
-      console.log('Checking user location...');
-      const newUserLocation = await Location.getCurrentPositionAsync({ accuracy: Location.Accuracy.Highest });
+  const locationInterval = () => {
+    if (trackingState === TrackingState.TRACKING) return 15000;
+    if (trackingState === TrackingState.TRACKINGPAUSED) return 5000;
+    return null;
+  };
 
-      const differentLatitude = newUserLocation.coords.latitude !== userLocation?.coords.latitude;
-      const differentLongitude = newUserLocation.coords.longitude !== userLocation?.coords.longitude;
-      if (differentLatitude || differentLongitude) updateUserLocation(newUserLocation);
-    },
-    trackingState === TrackingState.TRACKING ? 5000 : null,
-  );
+  useInterval(async () => {
+    const newUserLocation = await Location.getCurrentPositionAsync({ accuracy: Location.Accuracy.Highest });
+    const differentLatitude = newUserLocation.coords.latitude !== userLocation?.coords.latitude;
+    const differentLongitude = newUserLocation.coords.longitude !== userLocation?.coords.longitude;
+    if (differentLatitude || differentLongitude) updateUserLocation(newUserLocation);
+  }, locationInterval());
 
   const addUnUploadedActivity = (activity: Activities_Insert_Input) =>
     setUnUploadedActivities([...unUploadedActivities, activity]);
 
-  // Tracking
   const startTracking = () => {
-    if (insideGeoFence) {
-      setScore(0);
-      setDuration(0);
-      setTrackingState(TrackingState.TRACKING);
-      setTrackingStart(getCurrentTimestamp());
+    if (locationPermission && locationPermission.status !== 'granted') {
+      Alert.alert('Location permission not accepted', 'Location permission is required to start tracking');
+      return;
     }
+    if (!insideGeoFence) {
+      Alert.alert(
+        'Not inside a Hover zone',
+        "Sorry, you can't start tracking here! Move to a Hover zone to start earning points.",
+      );
+      return;
+    }
+
+    setScore(0);
+    setDuration(0);
+    setTrackingState(TrackingState.TRACKING);
+    setTrackingStart(getCurrentTimestamp());
+    setStartGeofence(insideGeoFence);
   };
   const resumeTracking = () => {
     setTrackingState(TrackingState.TRACKING);
@@ -179,6 +188,15 @@ export const TrackingProvider = ({ children }: Props) => {
   };
 
   const stopTracking = async (caption: string) => {
+    if (startGeoFence?.id !== insideGeoFence?.id) {
+      Alert.alert(
+        'Moved Hover zone during activity',
+        "You cannot end your activity in another Hover zone than where you started! Move back to: '" +
+          startGeoFence?.name +
+          "' to finish this activity. Or discard it...",
+      );
+      return;
+    }
     setTrackingState(TrackingState.EXPLORE);
     const activity: Activities_Insert_Input = {
       caption: caption,
