+++ theme/colors.ts
@@ -17,6 +17,7 @@ export const transparent = 'rgba(0, 0, 0, 0)';
 export const grayTransparent = 'rgba(66, 66, 66, 0.95)';
 export const greenTransparent = 'rgba(40, 167, 69, 0.95)';
 export const redTransparent = 'rgba(221, 46, 68, 0.95)';
+export const blueTransparent = 'rgba(3, 102, 214, 0.95)';
 export const almostBlackTransparent = 'rgba(18, 18, 18, 0.95)';
 
 // Base colors
