+++ types/hoverWithFriendsType.ts
@@ -1,0 +1,6 @@
+export enum HoverWithFriendState {
+  NONE,
+  STARTING,
+  JOINING,
+  ONGOING,
+}
